(#499) Move wait for pin tip into a common area
