+++ src/artemis/device_setup_plans/setup_oav.py
@@ -4,12 +4,13 @@ import bluesky.plan_stubs as bps
 import numpy as np
 from bluesky.utils import Msg
 from dodal.devices.oav.oav_calculations import camera_coordinates_to_xyz
-from dodal.devices.oav.oav_detector import OAV
+from dodal.devices.oav.oav_detector import MXSC, OAV
 from dodal.devices.oav.oav_errors import OAVError_ZoomLevelNotFound
 from dodal.devices.oav.oav_parameters import OAVParameters
 from dodal.devices.oav.utils import ColorMode, EdgeOutputArrayImageType
 from dodal.devices.smargon import Smargon
 
+from artemis.exceptions import WarningException
 from artemis.log import LOGGER
 
 Pixel = Tuple[int, int]
@@ -135,3 +136,19 @@ def get_move_required_so_that_beam_is_at_pixel(
         oav_params.micronsPerXPixel,
         oav_params.micronsPerYPixel,
     )
+
+
+def wait_for_tip_to_be_found(mxsc: MXSC):
+    pin_tip = mxsc.pin_tip
+    yield from bps.trigger(pin_tip, wait=True)
+    found_tip = yield from bps.rd(pin_tip)
+    if found_tip == pin_tip.INVALID_POSITION:
+        top_edge = yield from bps.rd(mxsc.top)
+        bottom_edge = yield from bps.rd(mxsc.bottom)
+        LOGGER.info(
+            f"No tip found with top/bottom of {list(top_edge), list(bottom_edge)}"
+        )
+        raise WarningException(
+            f"No pin found after {pin_tip.validity_timeout.get()} seconds"
+        )
+    return found_tip
